+++ src/test/java/com/woopaca/javachess/pieces/PieceTest.java
@@ -1,10 +1,10 @@
-package com.woopaca.javachess.chess.pieces;
+package com.woopaca.javachess.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static com.woopaca.javachess.chess.pieces.Piece.Type;
-import static com.woopaca.javachess.chess.pieces.Piece.Color;
+import static com.woopaca.javachess.pieces.Piece.Type;
+import static com.woopaca.javachess.pieces.Piece.Color;
 
 import static org.assertj.core.api.Assertions.assertThat;
